refactor: rename local-store filename to LocalStore
